requirements.txt
coverage
Django
djangorestframework
factory-boy
mock
coverage
django-timer
